add authentication middleware
