+++ 05-JWT-Basics/preferred/controllers/main.js
@@ -2,24 +2,7 @@ const jwt = require("jsonwebtoken");
 require("dotenv").config();
 
 const greeting = async (req, res) => {
-	const authHeader = req.headers.authorization;
-	console.log(authHeader)
-
-	if (!authHeader || !authHeader.startsWith("Bearer ")) {
-		console.log("An error occurred");
-		res.status(401).json({ messsage: "something went wrong" });
-	}
-	const token = authHeader.split(" ")[1];
-
-	try {
-		const encrypted = await jwt.verify(token, process.env.JWT_SECRET);
-		const {userName} = encrypted
-		// res.status(200).json({ message: encrypted });
-		res.status(200).send(`<h1>Hello and welcome ${userName} </h1>`);
-	} catch (error) {
-		console.log(error);
-		// res.status(400).json({message:"error"})
-	}
+	res.status(200).send(`<h1>Hello and welcome ${req.user.userName} </h1>`);
 };
 
 const logOn = async (req, res) => {
